Refactoring instance related classes to subfolder
