+++ tests/test_geometry.py
@@ -17,9 +17,16 @@ import unittest
 import pytest
 from parameterized import parameterized
 
-from wsidicom.geometry import (Orientation, Point, PointMm, Region, RegionMm,
-                               Size, SizeMm)
-from wsidicom.image_data import ImageOrigin
+from wsidicom.geometry import (
+    Orientation,
+    Point,
+    PointMm,
+    Region,
+    RegionMm,
+    Size,
+    SizeMm,
+)
+from wsidicom.instance import ImageOrigin
 
 
 @pytest.mark.unittest
@@ -60,108 +67,88 @@ class WsiDicomGeomtryTests(unittest.TestCase):
 
     def test_region_mm(self):
         region = RegionMm(PointMm(1.0, 2.0), SizeMm(3.0, 4.0))
-        self.assertEqual(
-            region.start,
-            PointMm(1.0, 2.0)
-        )
-        self.assertEqual(
-            region.end,
-            PointMm(4.0, 6.0)
-        )
+        self.assertEqual(region.start, PointMm(1.0, 2.0))
+        self.assertEqual(region.end, PointMm(4.0, 6.0))
 
     def test_region_mm_subtract(self):
         region = RegionMm(PointMm(1.0, 2.0), SizeMm(3.0, 4.0))
         region = region - PointMm(1.0, 2.0)
-        self.assertEqual(
-            region.start,
-            PointMm(0.0, 0.0)
-        )
-        self.assertEqual(
-            region.end,
-            PointMm(3.0, 4.0)
-        )
+        self.assertEqual(region.start, PointMm(0.0, 0.0))
+        self.assertEqual(region.end, PointMm(3.0, 4.0))
 
     def test_region_mm_add(self):
         region = RegionMm(PointMm(1.0, 2.0), SizeMm(3.0, 4.0))
         region = region + PointMm(1.0, 2.0)
-        self.assertEqual(
-            region.start,
-            PointMm(2.0, 4.0)
-        )
-        self.assertEqual(
-            region.end,
-            PointMm(5.0, 8.0)
-        )
+        self.assertEqual(region.start, PointMm(2.0, 4.0))
+        self.assertEqual(region.end, PointMm(5.0, 8.0))
 
-    @parameterized.expand([
-        (   # Image x along slide y, Image y along slide x
-            RegionMm(PointMm(2.0, 4.0), SizeMm(1.0, 2.0)),
-            ImageOrigin(PointMm(1.0, 2.0),  Orientation([0, 1, 0, 1, 0, 0])),
-            PointMm(2.0, 1.0),
-            PointMm(4.0, 2.0)
-        ),
-        (   # Image x reversed to slide y, Image y reversed to slide x
-            RegionMm(PointMm(1.0, 4.0), SizeMm(2.0, 1.0)),
-            ImageOrigin(PointMm(4.0, 8.0), Orientation([0, -1, 0, -1, 0, 0])),
-            PointMm(3.0, 1.0),
-            PointMm(4.0, 3.0)
-        ),
-        (   # Image x along slide x, Image y reversed to slide y
-            RegionMm(PointMm(2.0, 5.0), SizeMm(2.0, 1.0)),
-            ImageOrigin(PointMm(1.0, 8.0),  Orientation([1, 0, 0, 0, -1, 0])),
-            PointMm(1.0, 2.0),
-            PointMm(3.0, 3.0)
-        ),
-        (   # Image x reversed to slide x, Image y along slide y
-            RegionMm(PointMm(2.0, 3.0), SizeMm(2.0, 3.0)),
-            ImageOrigin(PointMm(5.0, 2.0), Orientation([-1, 0, 0, 0, 1, 0])),
-            PointMm(1.0, 1.0),
-            PointMm(3.0, 4.0)
-        )
-    ])
+    @parameterized.expand(
+        [
+            (  # Image x along slide y, Image y along slide x
+                RegionMm(PointMm(2.0, 4.0), SizeMm(1.0, 2.0)),
+                ImageOrigin(PointMm(1.0, 2.0), Orientation([0, 1, 0, 1, 0, 0])),
+                PointMm(2.0, 1.0),
+                PointMm(4.0, 2.0),
+            ),
+            (  # Image x reversed to slide y, Image y reversed to slide x
+                RegionMm(PointMm(1.0, 4.0), SizeMm(2.0, 1.0)),
+                ImageOrigin(PointMm(4.0, 8.0), Orientation([0, -1, 0, -1, 0, 0])),
+                PointMm(3.0, 1.0),
+                PointMm(4.0, 3.0),
+            ),
+            (  # Image x along slide x, Image y reversed to slide y
+                RegionMm(PointMm(2.0, 5.0), SizeMm(2.0, 1.0)),
+                ImageOrigin(PointMm(1.0, 8.0), Orientation([1, 0, 0, 0, -1, 0])),
+                PointMm(1.0, 2.0),
+                PointMm(3.0, 3.0),
+            ),
+            (  # Image x reversed to slide x, Image y along slide y
+                RegionMm(PointMm(2.0, 3.0), SizeMm(2.0, 3.0)),
+                ImageOrigin(PointMm(5.0, 2.0), Orientation([-1, 0, 0, 0, 1, 0])),
+                PointMm(1.0, 1.0),
+                PointMm(3.0, 4.0),
+            ),
+        ]
+    )
     def test_region_mm_to_other_origin(
         self,
         region: RegionMm,
         origin: ImageOrigin,
         expected_start: PointMm,
-        expected_end: PointMm
+        expected_end: PointMm,
     ):
         transformed_region = origin.transform_region(region)
         self.assertEqual(transformed_region.start, expected_start)
         self.assertEqual(transformed_region.end, expected_end)
 
-    @parameterized.expand([
-        (Region(Point(3, 4), Size(6, 4)), Point(9, 10), 2.0),
-        (Region(Point(9, 10), Size(6, 4)), Point(3, 4), 0.5),
-        (Region(Point(3, 4), Size(6, 4)), Point(15, 16), 3.0),
-        (Region(Point(4, 7), Size(2, 6)), Point(9, 17), 2.0),
-        (Region(Point(9, 17), Size(2, 6)), Point(4, 7), 0.5),
-        (Region(Point(4, 7), Size(2, 6)), Point(14, 27), 3.0),
-    ])
-    def test_region_zoom(
-        self,
-        region: Region,
-        expected_start: Point,
-        zoom: float
-    ):
+    @parameterized.expand(
+        [
+            (Region(Point(3, 4), Size(6, 4)), Point(9, 10), 2.0),
+            (Region(Point(9, 10), Size(6, 4)), Point(3, 4), 0.5),
+            (Region(Point(3, 4), Size(6, 4)), Point(15, 16), 3.0),
+            (Region(Point(4, 7), Size(2, 6)), Point(9, 17), 2.0),
+            (Region(Point(9, 17), Size(2, 6)), Point(4, 7), 0.5),
+            (Region(Point(4, 7), Size(2, 6)), Point(14, 27), 3.0),
+        ]
+    )
+    def test_region_zoom(self, region: Region, expected_start: Point, zoom: float):
         zoomed_region = region.zoom(zoom)
 
         self.assertEqual(zoomed_region.start, expected_start)
         self.assertEqual(zoomed_region.size, region.size)
 
-    @parameterized.expand([
-        (RegionMm(PointMm(3, 4), SizeMm(6, 4)), PointMm(9, 10), 2.0),
-        (RegionMm(PointMm(9, 10), SizeMm(6, 4)), PointMm(3, 4), 0.5),
-        (RegionMm(PointMm(3, 4), SizeMm(6, 4)), PointMm(15, 16), 3.0),
-        (RegionMm(PointMm(4, 7), SizeMm(2, 6)), PointMm(9, 17), 2.0),
-        (RegionMm(PointMm(9, 17), SizeMm(2, 6)), PointMm(4, 7), 0.5),
-        (RegionMm(PointMm(4, 7), SizeMm(2, 6)), PointMm(14, 27), 3.0),
-    ])
+    @parameterized.expand(
+        [
+            (RegionMm(PointMm(3, 4), SizeMm(6, 4)), PointMm(9, 10), 2.0),
+            (RegionMm(PointMm(9, 10), SizeMm(6, 4)), PointMm(3, 4), 0.5),
+            (RegionMm(PointMm(3, 4), SizeMm(6, 4)), PointMm(15, 16), 3.0),
+            (RegionMm(PointMm(4, 7), SizeMm(2, 6)), PointMm(9, 17), 2.0),
+            (RegionMm(PointMm(9, 17), SizeMm(2, 6)), PointMm(4, 7), 0.5),
+            (RegionMm(PointMm(4, 7), SizeMm(2, 6)), PointMm(14, 27), 3.0),
+        ]
+    )
     def test_region_mm_zoom(
-        self,
-        region: RegionMm,
-        expected_start: PointMm,
-        zoom: float
+        self, region: RegionMm, expected_start: PointMm, zoom: float
     ):
         zoomed_region = region.zoom(zoom)
