+++ wsidicom/instance/tile_index/sparse_tile_index.py
@@ -19,7 +19,7 @@ from pydicom.dataset import Dataset
 
 from wsidicom.errors import WsiDicomNotFoundError
 from wsidicom.geometry import Point, Size
-from wsidicom.tile_index.tile_index import TileIndex
+from wsidicom.instance.tile_index.tile_index import TileIndex
 
 
 class SparseTilePlane:
