+++ wsidicom/wsidicom_file/wsidicom_file.py
@@ -25,10 +25,10 @@ from pydicom.tag import BaseTag, ItemTag, SequenceDelimiterTag, Tag
 from pydicom.uid import UID
 
 from wsidicom.config import settings
-from wsidicom.dataset import ImageType, WsiDataset
 from wsidicom.errors import WsiDicomFileError
 from wsidicom.file import OffsetTableType, WsiDicomFileBase
 from wsidicom.geometry import Size
+from wsidicom.instance import ImageType, WsiDataset
 from wsidicom.uid import FileUids, SlideUids
