+++ wsidicom/wsidicom_file/wsidicom_file_image_data.py
@@ -18,8 +18,8 @@ from typing import List, OrderedDict, Sequence, Union
 from pydicom.uid import UID
 
 from wsidicom.errors import WsiDicomNotFoundError
+from wsidicom.instance import WsiDicomImageData
 from wsidicom.wsidicom_file.wsidicom_file import WsiDicomFile
-from wsidicom.image_data import WsiDicomImageData
 
 
 class WsiDicomFileImageData(WsiDicomImageData):
